backend/pyproject.toml
[project]
name = "acom-backend"
version = "0.1.0"
description = ""
authors = [
    {name = "adminBR",email = "luis.cardosodev@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "djangorestframework (>=3.16.1,<4.0.0)",
    "django-cors-headers (>=4.9.0,<5.0.0)",
    "uvicorn (>=0.35.0,<0.36.0)",
    "redis (>=6.4.0,<7.0.0)",
    "psycopg2-binary (>=2.9.10,<3.0.0)",
    "celery (>=5.5.3,<6.0.0)",
    "pydantic (>=2.11.7,<3.0.0)",
    "python-telegram-bot (>=22.3,<23.0)",
    "python-dotenv (>=1.1.1,<2.0.0)",
    "whitenoise (>=6.9.0,<7.0.0)",
    "gunicorn (>=23.0.0,<24.0.0)",
    "drf-spectacular[sidecar] (>=0.28.0,<0.29.0)",
    "requests (>=2.32.5,<3.0.0)",
]
[tool.poetry]
package-mode = false

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
fakeredis = "^2.31.1"

